Init content placeholder block
