+++ scripts/scripts.js
@@ -1355,11 +1355,6 @@ async function loadPage() {
   showBrowseBackgroundGraphic();
   showSignupDialog();
 
-  if (isProfilePage()) {
-    await loadDefaultModule(`${window.hlx.codeBasePath}/scripts/profile/personalized-home.js`);
-    document.body.classList.remove('loading');
-  }
-
   if (isDocArticlePage()) {
     // wrap main content in a div - UGP-11165
     const main = document.querySelector('main');
@@ -1394,7 +1389,6 @@ if (!window.hlx.DO_NOT_LOAD_PAGE) {
   const { lang } = getPathDetails();
   document.documentElement.lang = lang || 'en';
   if (isProfilePage()) {
-    document.body.classList.add('loading');
     if (window.location.href.includes('.html')) {
       loadPage();
     } else {
